requirements.txt
# Core AutoGen and LLM dependencies (Python 3.13 compatible)
autogen-agentchat>=0.2.0
openai>=1.0.0

# Financial data APIs
yfinance>=0.2.32
finnhub-python>=2.4.19
sec-edgar-downloader==5.0.3

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# RAG and embeddings for 10-K analysis
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
chromadb>=0.4.0
tiktoken>=0.5.0

# PDF and document processing
pypdf>=3.17.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# HTTP requests
requests>=2.31.0
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0

# Logging and monitoring (optional - for RAG evaluation)
arize-phoenix>=2.0.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.0
